NTP-607: Implement search tuition partners by name
