+++ UI/Pages/FullList.cshtml.cs
@@ -1,7 +1,9 @@
 using Application.Handlers;
 using Domain.Search;
 using MediatR;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using UI.Extensions;
 
 namespace UI.Pages;
 
@@ -11,12 +13,19 @@ public class FullList : PageModel
 
     public FullList(IMediator mediator) => _mediator = mediator;
 
+    [BindProperty(SupportsGet = true)]
+    public SearchModel Data { get; set; } = new();
+
     public TuitionPartnerSearchResultsPage? Results { get; private set; }
 
     public async Task OnGet()
     {
+        Data.From = ReferrerList.FullList;
+        TempData.Set("AllSearchData", Data);
+
         Results = await _mediator.Send(new SearchTuitionPartnerHandler.Command
         {
+            Name = Data.Name,
             OrderBy = TuitionPartnerOrderBy.Name
         });
     }
